[Feat] Add Store Map API
